moved fetchPublicJsonAsset from helpers to datastore/datastore=helpers
