+++ functions/datastore/appointments.js
@@ -12,8 +12,8 @@ const PROTOTYPE = '/datastore/appointment-prototype.json';
 const FHIR_APPOINTMENT = 'Appointments';
 
 const assert = require("assert");
-const { getParam, fetchPublicJsonAsset } = require(Runtime.getFunctions()['helpers'].path);
-const { read_fhir, save_fhir } = require(Runtime.getFunctions()['datastore/datastore-helpers'].path);
+const { getParam } = require(Runtime.getFunctions()['helpers'].path);
+const { read_fhir, save_fhir, fetchPublicJsonAsset } = require(Runtime.getFunctions()['datastore/datastore-helpers'].path);
 
 // --------------------------------------------------------------------------------
 function transform_fhir_to_appointment(fhir_appointment) {
