+++ functions/datastore/datastore-helpers.private.js
@@ -9,6 +9,8 @@
  */
 
 const assert = require('assert');
+const http = require("http");
+const https = require("https");
 
 /*
  * ----------------------------------------------------------------------------------------------------
@@ -219,6 +221,48 @@ async function save_fhir(context, syncServiceSid, resourceType, resources) {
   return document ? document.sid : null;
 }
 
+
+/*
+ * ----------------------------------------------------------------------------------------------------
+ * fetches content of public json asset via https
+ *
+ * parameters
+ * - context: Twilio runtime context
+ * - assetPath: url path to asset
+ *
+ * returns: json content of asset
+ * ----------------------------------------------------------------------------------------------------
+ */
+async function fetchPublicJsonAsset(context, assetPath) {
+  const hostname = context.DOMAIN_NAME.split(':')[0];
+  const port = context.DOMAIN_NAME.split(':')[1];
+  const options = {
+    hostname: hostname,
+    port: port,
+    path: assetPath,
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+  };
+  const http_protocol = (hostname === 'localhost') ? http : https;
+
+  return new Promise((resolve, reject) => {
+    const request = http_protocol.request(options, (response) => {
+      let data = '';
+      response.on('data', (chunk) => {
+        data += chunk;
+      });
+      response.on('end', () => {
+        resolve(JSON.parse(data));
+      });
+      response.on('error', (error) => {
+        reject(error);
+      });
+    });
+    request.end();
+  });
+}
+
+
 // --------------------------------------------------------------------------------
 module.exports = {
   read_fhir,
@@ -227,5 +271,6 @@ module.exports = {
   upsertSyncDocument,
   deleteSyncDocument,
   fetchSyncMapItem,
-  insertSyncMapItem
+  insertSyncMapItem,
+  fetchPublicJsonAsset,
 };
