+++ functions/datastore/providers.js
@@ -13,8 +13,8 @@ const FHIR_PRACTITIONER = 'Practitioners';
 const FHIR_PRACTITIONER_ROLE = 'PractitionerRoles';
 
 const assert = require("assert");
-const { getParam, fetchPublicJsonAsset } = require(Runtime.getFunctions()['helpers'].path);
-const { read_fhir } = require(Runtime.getFunctions()['datastore/datastore-helpers'].path);
+const { getParam } = require(Runtime.getFunctions()['helpers'].path);
+const { read_fhir, save_fhir, fetchPublicJsonAsset } = require(Runtime.getFunctions()['datastore/datastore-helpers'].path);
 
 // --------------------------------------------------------------------------------
 function transform_fhir_to_provider(fhir_practitioner, fhir_practitioner_roles) {
