+++ functions/helpers.private.js
@@ -196,125 +196,9 @@ async function getParam(context, key) {
   }
 }
 
-/*
- * --------------------------------------------------------------------------------
- * get/add/remove flow friendly name
- *
- * use environment variable TWILIO_FLOWS
- * --------------------------------------------------------------------------------
- */
-function getFlowFriendlyNames(context) {
-  const flows = JSON.parse(context.APPLICATION_STUDIO_FLOWS);
-  return flows;
-}
-
-// --------------------------------------------------------------------------------
-function addFlowFriendlyName(context, friendlyName) {
-  const flows = JSON.parse(context.APPLICATION_STUDIO_FLOWS);
-  if (! friendlyName) {
-    console.log('friendlyName is empty!');
-    return null;
-  }
-
-  try {
-    if (flows.indexOf(friendlyName) > -1) {
-      console.log(`friendlyName=${friendlyName} already exists!`);
-      return null;
-    }
-    flows.push(friendlyName);
-
-    setParam(context, 'APPLICATION_STUDIO_FLOWS', JSON.stringify(flows));
-  } catch (err) {
-    console.log(`error setting friendlyName=${friendlyName}!`);
-    return null;
-  }
-
-  return friendlyName;
-}
-
-// --------------------------------------------------------------------------------
-function removeFlowFriendlyName(context, friendlyName) {
-  const flows = JSON.parse(context.APPLICATION_STUDIO_FLOWS);
-
-  if (! friendlyName) {
-    console.log('friendlyName is empty!');
-    return null;
-  }
-
-  try {
-    i = flows.indexOf(friendlyName);
-    if (i === -1) {
-      console.log(`friendlyName=${friendlyName} does not exist!`);
-      return null;
-    }
-    flows.splice(friendlyName, 1);
-
-    setParam(context, 'APPLICATION_STUDIO_FLOWS', JSON.stringify(flows));
-  } catch (err) {
-    console.log(`error setting friendlyName=${friendlyName}!`);
-    return null;
-  }
-
-  return friendlyName;
-}
-
-// --------------------------------------------------------------------------------
-async function assignPhone2Flow(context, flow_sid) {
-  const PHONE_NUMBER = await getParam(context, 'TWILIO_PHONE_NUMBER');
-
-  const client = context.getTwilioClient();
-
-  const flow = await client.studio.flows(flow_sid).fetch();
-
-  const phoneList = await client.incomingPhoneNumbers.list();
-  const phone = phoneList.find(p => p.phoneNumber === PHONE_NUMBER);
-  assert(phone, `Phone number ${PHONE_NUMBER} not found!!!`);
-
-  await client.incomingPhoneNumbers(phone.sid).update({
-    smsUrl: flow.webhookUrl
-  });
-
-  return phone.sid;
-}
-
-// --------------------------------------------------------------------------------
-async function fetchPublicJsonAsset(context, assetPath) {
-  const hostname = context.DOMAIN_NAME.split(':')[0];
-  const port = context.DOMAIN_NAME.split(':')[1];
-  const options = {
-    hostname: hostname,
-    port: port,
-    path: assetPath,
-    method: 'GET',
-    headers: { 'Content-Type': 'application/json' },
-  };
-  const http_protocol = (hostname === 'localhost') ? http : https;
-
-  return new Promise((resolve, reject) => {
-    const request = http_protocol.request(options, (response) => {
-      let data = '';
-      response.on('data', (chunk) => {
-        data += chunk;
-      });
-      response.on('end', () => {
-        resolve(JSON.parse(data));
-      });
-      response.on('error', (error) => {
-        reject(error);
-      });
-    });
-    request.end();
-  });
-}
-
 
 // --------------------------------------------------------------------------------
 module.exports = {
   getParam,
   setParam,
-  getFlowFriendlyNames,
-  addFlowFriendlyName,
-  removeFlowFriendlyName,
-  assignPhone2Flow,
-  fetchPublicJsonAsset,
 };
